Refactor: Populate course location options dynamically
This commit refactors the way course location options are populated in the Tutor LMS course builder. Instead of hardcoding the options, the `custom-fields.js` script now dynamically generates the options based on the `tlcfData.locations` array passed from PHP. The AJAX call and initial console logs are commented out, and a temporary console log is added for debugging purposes. A `foreach` loop is used to iterate through the locations and create an array of options with label and value properties. This allows for a more flexible and maintainable way to manage course location options.