Fix: Correct locations data source and update script version
The locations data source in `custom-fields.js` was incorrectly referencing `tlcfData.values` instead of `tlcfData.locations`. This commit corrects the reference to ensure the correct data is used.

Additionally, the version number of the `tlcf-custom-fields-js` script is updated from 1.0.2 to 1.0.3 in the `functions.php` file. This ensures that the updated script is loaded correctly, especially after changes or fixes have been applied. This helps to prevent caching issues and ensures users get the latest version of the script.