Fix: Correct locations data source and update script version
The locations data source in `custom-fields.js` was incorrectly referencing `tlcfData.locations` instead of `tlcfData.values`. This commit corrects the reference to ensure the correct data is used. The `locations` variable is now assigned the value of `tlcfData.values`.

Additionally, the version number of the `tlcf-custom-fields-js` script is updated from 1.0.3 to 1.0.4 in the `functions.php` file. This ensures that the updated script is loaded correctly, especially after changes or fixes have been applied. This helps to prevent caching issues and ensures users get the latest version of the script. The `socio_connect_locations` option is now passed as an array of values to the `tlcfData` variable. The `implode` function was removed.